【绩效等级后端计算】新增和更改时自动计算绩效等级【2.1】

+++ src/main/java/com/aioveu/boot/aioveuPerformance/model/form/AioveuPerformanceForm.java
@@ -36,7 +36,8 @@ public class AioveuPerformanceForm implements Serializable {
 
     @Schema(description = "KPI评分(1-100分)")
     @NotNull(message = "KPI评分(1-100分)不能为空")
-    private Integer kpiScore;
+    // 将Integer改为BigDecimal
+    private BigDecimal kpiScore;
 
     @Schema(description = "生产率百分比(%)")
     @NotNull(message = "生产率百分比(%)不能为空")
